unnecessary int cast removed

+++ cores/arduino/wiring_pulse.cpp
@@ -164,7 +164,7 @@ unsigned long pulseIn(PinName pin, PinStatus state, unsigned long timeout)
         while (!pulseSecond && (micros() - startMicros < timeout) ) {
             pulseSecond = nrf_timer_cc_read(PULSE_TIMER, TIMER_SECOND_CHANNEL);
         }
-        pulseTime = pulseSecond ? (uint32_t) ( (int)pulseSecond - (int)pulseFirst) : TIMEOUT_US;
+        pulseTime = pulseSecond ? pulseSecond - pulseFirst : TIMEOUT_US;
     }
     
     /* Deallocate all the PPI channels, events and groups */
